layout for playlist created
