+++ skypro-music/src/app/layout.tsx
@@ -25,3 +25,5 @@ export default function RootLayout({
     </html>
   );
 }
+
+
