+++ skypro-music/src/components/CenterBlock/CenterBlock.module.css
@@ -1,10 +1,10 @@
-.main {
+/* .main {
     width: auto;
     -webkit-box-flex: 3;
     -ms-flex-positive: 3;
     flex-grow: 3;
     padding: 20px 40px 20px 111px;
-}
+} */
 
 .title {
     font-style: normal;
