--- skypro-music/src/components/Main/Main.module.css
@@ -1,18 +1,0 @@
-.main {
-  -webkit-box-flex: 1;
-  -ms-flex: 1 1 auto;
-  flex: 1 1 auto;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -ms-flex-wrap: wrap;
-  flex-wrap: wrap;
-  -webkit-box-pack: justify;
-  -ms-flex-pack: justify;
-  justify-content: space-between;
-  height: 100%;
-  overflow: hidden;
-  overflow-y: auto;
-  scrollbar-color: #4E4E4E #000000;
-  scrollbar-width: thin;
-}
